+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - All JavaScript and CSS files found in `config/diy-widget` are now loaded for CP requests, not just the ones matching template filenames.
+- It’s now possible to register template hooks by uploading PHP files to `config/diy-widget`. ([#9](https://github.com/carlcs/craft-diywidget/issues/9))
 - Widget titles are now displayed in Title Case.
 
 ## 2.0.2 - 2020-04-02